combine prediction_dataset.py & dataset.py
